pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "arweave-today-podcaster"
version = "1.0.0"
description = "An intelligent, automated podcast generator for Arweave ecosystem news"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Arweave Ecosystem", email = "contact@arweave.org"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "yt-dlp>=2024.7.16",
    "faster-whisper>=1.0.3",
    "google-generativeai",
    "urllib3",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "isort>=5.0",
    "flake8>=3.8",
    "mypy>=0.800",
]

[project.scripts]
arweave-podcaster = "arweave_podcaster.core.podcast_generator:main"

[project.urls]
Homepage = "https://github.com/yourusername/arweave-today-ai-podcaster"
Repository = "https://github.com/yourusername/arweave-today-ai-podcaster.git"
Documentation = "https://github.com/yourusername/arweave-today-ai-podcaster#readme"
"Bug Tracker" = "https://github.com/yourusername/arweave-today-ai-podcaster/issues"

[tool.setuptools]
packages = ["arweave_podcaster"]

[tool.setuptools.package-data]
arweave_podcaster = ["*.md", "*.txt"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["arweave_podcaster"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
